% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ARMI_markers.r
\name{addARMIMarkers}
\alias{addARMIMarkers}
\title{This function adds ARMI markers to the map with popups containing ggplot graphs.}
\usage{
addARMIMarkers(mapProxy, data, riverflyARMIData, screen_width)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{data}{A data frame containing the site-averaged recent ARMI data to be plotted.}

\item{riverflyARMIData}{A data frame containing all ARMI data for generating the ggplot graphs.}

\item{input}{The Shiny input object to access screen width for responsive popup sizing.}
}
\description{
This function adds ARMI markers to the map with popups containing ggplot graphs.
}
