% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_invasive_species_markers.r
\name{addInvasiveSpeciesMarkers}
\alias{addInvasiveSpeciesMarkers}
\title{This function adds Invasive species markers to the map.}
\usage{
addInvasiveSpeciesMarkers(mapProxy, data, invasiveType, plot_palette)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{data}{A data frame containing the Invasive species data to be plotted.}

\item{invasiveType}{The specific Invasive species to filter and plot.}

\item{plot_palette}{A vector of colors to use for the invasive species markers.}
}
\description{
This function adds Invasive species markers to the map.
}
