% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_other_species_markers.r
\name{addOtherSpeciesMarkers}
\alias{addOtherSpeciesMarkers}
\title{This function adds Other species markers to the map.}
\usage{
addOtherSpeciesMarkers(mapProxy, fullData, otherSpecies)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{fullData}{A data frame containing the Other species data to be plotted.}

\item{otherSpecies}{The specific Other species to filter and plot.}
}
\description{
This function adds Other species markers to the map.
}
