% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_polygons_and_lines.r
\name{addPolygonsAndLines}
\alias{addPolygonsAndLines}
\title{This function adds polygons and lines to the map based on the zoom level.}
\usage{
addPolygonsAndLines(mapProxy, zoomLevel)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{zoomLevel}{The current zoom level of the map.}
}
\description{
This function adds polygons and lines to the map based on the zoom level.
}
