% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_water_quality_markers.r
\name{addWaterQualityMarkers}
\alias{addWaterQualityMarkers}
\title{Add water quality markers to a Leaflet map proxy}
\usage{
addWaterQualityMarkers(mapProxy, wq_data_recent, wq_data, metric, screen_width)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{wq_data_recent}{A data frame containing recent average water quality data for generating the ggplot graphs.}

\item{wq_data}{A list containing two data frames: all observations and recent averages.}

\item{input}{Shiny input object for reactive inputs.}
}
\description{
This function adds water quality markers to a Leaflet map proxy with popups containing ggplot graphs.
}
