% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_helpers.R
\name{anonymise_organisations}
\alias{anonymise_organisations}
\title{Anonymise organisations based on sign-up sheet preferences}
\usage{
anonymise_organisations(data_frame)
}
\arguments{
\item{data_frame}{A data frame containing an 'organisation' column.}
}
\value{
A data frame with specified organisations anonymised.
}
\description{
Anonymise organisations based on sign-up sheet preferences
}
