% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{choices_list}
\alias{choices_list}
\title{Text for the choices for water quality, outfall, and abundance form inputs.}
\usage{
choices_list
}
\description{
Text for the choices for water quality, outfall, and abundance form inputs.
}
