% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_helpers.R
\name{flip_site_names}
\alias{flip_site_names}
\title{Flip site names from "Site, Organisation" to "Organisation, Site"}
\usage{
flip_site_names(site_name)
}
\arguments{
\item{site_name}{A character string representing the site name.}
}
\value{
A character string with the site name flipped.
}
\description{
Flip site names from "Site, Organisation" to "Organisation, Site"
}
