% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{isValidEmail}
\alias{isValidEmail}
\title{Check for a valid email address}
\usage{
isValidEmail(x)
}
\arguments{
\item{x}{entered string in the email address field}
}
\value{
boolean whether email address format is correct
}
\description{
Check for a valid email address
}
