% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{label_mandatory}
\alias{label_mandatory}
\title{Mimics Google form functionality for setting required fields}
\usage{
label_mandatory(label)
}
\arguments{
\item{label}{the label of the required form input}
}
\description{
Mimics Google form functionality for setting required fields
}
