% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_fct_process_data.R
\name{make_ARMI_plot_data}
\alias{make_ARMI_plot_data}
\title{make_ARMI_plot_data}
\usage{
make_ARMI_plot_data(Riverfly_ARMI_Calc, Unique_BRC_Sampling_Locs)
}
\arguments{
\item{Riverfly_ARMI_Calc}{A dataframe output from make_riverfly_ARMI that contains
riverfly data and the ARMI column}

\item{Unique_BRC_Sampling_Locs}{a data frame that contains the geographic coordinates
for the sampling sites.}
}
\value{
averaged site ARMI calculations to plot on the riverfly figure
}
\description{
A function to make the plot object for the ARMI map and tooltip graphs
}
