% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recent_inv_spp.R
\name{make_recent_inv_spp}
\alias{make_recent_inv_spp}
\title{make_recent_inv_spp}
\usage{
make_recent_inv_spp(cleaned_data, sampling_locs, plot_palette)
}
\arguments{
\item{cleaned_data}{The cleaned invasive species data}

\item{sampling_locs}{The sampling locations for invasive species data}

\item{plot_palette}{The colour palette to use for the plot}
}
\value{
The data frame for the recent invasive species plot.
}
\description{
This function makes the data for the recent invasive species plot
}
