% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_fct_make_ARMI.R
\name{make_riverfly_ARMI}
\alias{make_riverfly_ARMI}
\title{make_riverfly_ARMI}
\usage{
make_riverfly_ARMI(table_name)
}
\arguments{
\item{table_name}{- name of the table to extract the riverfly data from}
}
\value{
The data frame with riverfly data and an appended ARMI column.
}
\description{
A function to add the ARMI values to the riverfly data
}
