% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_water_quality_plot_data.r
\name{make_water_quality_plot_data}
\alias{make_water_quality_plot_data}
\title{make_water_quality_plot_data}
\usage{
make_water_quality_plot_data(water_quality_data, sampling_locs)
}
\arguments{
\item{water_quality_data}{The cleaned water quality data}

\item{sampling_locs}{The sampling locations for water quality data}

\item{reading_type}{The type of water quality reading (e.g., "Temperature (\u00B0C)")}
}
\value{
A data frame for the water quality plot
}
\description{
Function to create the data for the Water Quality plots
}
