% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{plot_breaks}
\alias{plot_breaks}
\title{plot_breaks}
\usage{
plot_breaks
}
\description{
These are the breaks used for plotting water quality and ARMI data.
They define the bins for categorizing readings into different levels.
}
