% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_populate_db.R
\name{populate_db}
\alias{populate_db}
\title{populate_db}
\usage{
populate_db(data_to_insert, table_name)
}
\arguments{
\item{data_to_insert}{A data frame containing the data to be inserted into the database.}

\item{table_name}{Name of the SQLite table where data should be inserted.}
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
