% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_put_table_data.R
\name{put_table_data}
\alias{put_table_data}
\title{Function to get the tables from the new Google Sheet}
\usage{
put_table_data(
  sheet_url,
  data_types = c("riverfly", "water_quality", "outfall_safari", "invasive_species")
)
}
\arguments{
\item{sheet_url}{The URL of the Google Sheet containing the data.}
}
\value{
A list of data frames, each corresponding to a table in the Google Sheet.
}
\description{
Function to get the tables from the new Google Sheet
}
