% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{remove_parenthesised_orgs}
\alias{remove_parenthesised_orgs}
\title{Remove parenthesised organisations from site names}
\usage{
remove_parenthesised_orgs(data_frame)
}
\arguments{
\item{data_frame}{A data frame containing a 'sampling_site' column.}
}
\value{
A data frame with parenthesised organisations removed from site names.
}
\description{
Remove parenthesised organisations from site names
}
