% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{riverfly_spp_bw}
\alias{riverfly_spp_bw}
\title{Riverfly species and water quality reading type mappings
A set of named lists that map internal column names to human-readable names for Riverfly species and water quality reading types.
These mappings are used throughout the package for data processing and visualization.}
\format{
A list of named lists:
\describe{
\item{riverfly_spp_bw}{A named list mapping Riverfly species internal column names to human-readable names.}
\item{other_spp_bw}{A named list mapping other species internal column names to human-readable names.}
\item{water_quality_bw}{A named list mapping water quality reading type internal column
names to human-readable names.}
}
}
\usage{
riverfly_spp_bw
}
\description{
Riverfly species and water quality reading type mappings
A set of named lists that map internal column names to human-readable names for Riverfly species and water quality reading types.
These mappings are used throughout the package for data processing and visualization.
}
\keyword{datasets}
