% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{subset_data}
\usage{
subset_data(input_df, col_name_start, col_name_end, data_type_name)
}
\arguments{
\item{input_df}{Data frame containing the raw data from Google Sheets.}

\item{col_name_start}{The starting column name for selecting relevant data.}

\item{col_name_end}{The ending column name for selecting relevant data.}

\item{data_type_name}{A string indicating the type of data being processed (e.g
"Urban Riverfly", "Water Quality", etc.) for warning messages.}
}
\value{
A cleaned data frame ready for analysis.
}
\description{
#' Function to subset data from Google Sheets for the River Champions project
}
\details{
This function imports data from Google Sheets and breaks it into several different data types
}
