% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_up_ARMI.r
\name{sum_up_ARMI}
\alias{sum_up_ARMI}
\title{Sum up ARMI observations across taxa}
\usage{
sum_up_ARMI(Riverfly_ARMI)
}
\arguments{
\item{Riverfly_ARMI}{A data frame containing Riverfly ARMI observations per taxa}
}
\value{
A data frame with ARMI summed across taxa
}
\description{
Sum up ARMI observations across taxa
}
