% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_turn_newsheet_into_db.R
\name{turn_newsheet_into_db}
\alias{turn_newsheet_into_db}
\title{Function to read in data from Google Sheets and populate the SQLite database
Also saves cleaned data as internal package data}
\usage{
turn_newsheet_into_db(
  data_types = c("Urban Riverfly", "Water Quality", "Urban Outfall Safari",
    "Invasive Species"),
  table_name = c("riverfly", "water_quality", "outfall_safari", "invasive_species"),
  col_indices = c(6, 6, 7, 7)
)
}
\arguments{
\item{data_types}{A character vector specifying which data types to process.}

\item{table_name}{A character vector specifying the corresponding SQLite table names for each data type.}

\item{col_indices}{A numeric vector specifying the starting column index for each data type in the Google Sheet.
Options include "riverfly", "water_quality", "invasive_species", and "
outfall_safari". Default is to process all data types.}
}
\value{
None. The function creates/updates the SQLite database and saves cleaned data as internal package data.
}
\description{
Function to read in data from Google Sheets and populate the SQLite database
Also saves cleaned data as internal package data
}
