% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{wq_y_axes}
\alias{wq_y_axes}
\title{List that includes the y-axis for popupGraphs for the water chemistry plots}
\format{
An object of class \code{list} of length 5.
}
\usage{
wq_y_axes
}
\description{
List that includes the y-axis for popupGraphs for the water chemistry plots
}
\keyword{datasets}
