% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ARMI_markers.r
\name{addARMIMarkers}
\alias{addARMIMarkers}
\title{This function adds ARMI markers to the map with popups containing ggplot graphs.}
\usage{
addARMIMarkers(mapProxy, map_data, popup_data, screen_width)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{map_data}{A data frame containing the site-averaged recent ARMI data to be plotted on the map.}

\item{popup_data}{A data frame containing all ARMI data for generating the ggplot graphs.}

\item{input}{The Shiny input object to access screen width for responsive popup sizing.}
}
\description{
This function adds ARMI markers to the map with popups containing ggplot graphs.
}
