% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_riverfly_species_markers.r
\name{addRiverflySpeciesMarkers}
\alias{addRiverflySpeciesMarkers}
\title{This function adds Urban Riverfly species markers to the map with popups containing ggplot graphs.}
\usage{
addRiverflySpeciesMarkers(
  mapProxy,
  map_data,
  popup_data,
  taxaType,
  screen_width
)
}
\arguments{
\item{mapProxy}{A leaflet map proxy object.}

\item{map_data}{A data frame containing the Urban Riverfly species data to be plotted on the map.}

\item{popup_data}{A data frame containing all Urban Riverfly species data for generating the ggplot graphs.}

\item{taxaType}{The specific Urban Riverfly species to filter and plot.}

\item{input}{The Shiny input object to access screen width for responsive popup sizing.}
}
\description{
This function adds Urban Riverfly species markers to the map with popups containing ggplot graphs.
}
