% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{choices_list}
\alias{choices_list}
\title{Text for the choices for water quality, outfall, and abundance form inputs.}
\format{
An object of class \code{list} of length 8.
}
\usage{
choices_list
}
\description{
Text for the choices for water quality, outfall, and abundance form inputs.
}
\keyword{datasets}
