% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(
  input_df,
  col_name_start,
  col_name_end,
  sample_site,
  locations_name,
  data_type_name
)
}
\arguments{
\item{input_df}{Data frame containing the raw data from Google Sheets.}

\item{col_name_start}{The starting column name for selecting relevant data.}

\item{col_name_end}{The ending column name for selecting relevant data.}

\item{sample_site}{Column in input_df that is used to filter out data uploads for this specific data type.}

\item{locations_name}{The name of the locations table in the SQLite database.}

\item{data_type_name}{A string indicating the type of data being processed (e.g
"Urban Riverfly", "Water Quality", etc.) for warning messages.}
}
\value{
A cleaned data frame ready for analysis.
}
\description{
#' Function to clean and process data from Google Sheets for the River Champions project
}
\details{
This function imports data from Google Sheets, breaks it into several different data types,
and processes it to prepare for analysis and visualization.
}
