% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_submission_df.R
\name{create_blank_submission_df}
\alias{create_blank_submission_df}
\title{Create empty submissions dataframe}
\usage{
create_blank_submission_df()
}
\value{
A tibble with all submission columns defined as character vectors.
}
\description{
Create empty submissions dataframe
}
