% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_pop_db.r
\name{db_create_and_pop}
\alias{db_create_and_pop}
\title{db_create
A function to create a SQLite database with specified tables for storing Riverfly and Water Quality data.}
\usage{
db_create_and_pop(sub_table, index_of_site_col, table_name, ...)
}
\arguments{
\item{sub_table}{The subsetted data frame read in from the Google Sheet.}

\item{index_of_site_col}{The index of the column in the cleaned data that contains the sampling site information.}

\item{table_name}{Name of the SQLite table file to create.}

\item{...}{Additional arguments to pass to the db_create function.}
}
\value{
The return value, if any, from executing the function.
}
\description{
db_create
A function to create a SQLite database with specified tables for storing Riverfly and Water Quality data.
}
