% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_pop_db.r
\name{db_create_and_pop}
\alias{db_create_and_pop}
\title{db_create
A function to create a SQLite database with specified tables for storing Riverfly and Water Quality data.}
\usage{
db_create_and_pop(
  full_form,
  locations_list,
  data_type,
  index_of_site_col,
  table_name,
  ...
)
}
\arguments{
\item{data_type}{A string indicating the type of data being processed (e.g., "Urban Riverfly", "Water Quality", etc.).}

\item{index_of_site_col}{The index of the column in the cleaned data that contains the sampling site information.}

\item{table_name}{Name of the SQLite table file to create.}

\item{BRC_full}{The full data frame read in from the Google Sheet.}
}
\value{
The return value, if any, from executing the function.
}
\description{
db_create
A function to create a SQLite database with specified tables for storing Riverfly and Water Quality data.
}
