% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_card.R
\name{information_card}
\alias{information_card}
\title{This function creates information cards on the page for a given
set of text, image, and links}
\usage{
information_card(title, description, alt_text = NULL, url, image_src, ns)
}
\arguments{
\item{title}{The title of the link}

\item{description}{The description underneath the link}

\item{alt_text}{Alternative text for the provided image, defaults to NULL}

\item{url}{The URL that the header link should target}

\item{image_src}{The location within the app of the image}
}
\description{
This function creates information cards on the page for a given
set of text, image, and links
}
