% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_fct_process_data.R
\name{make_riverfly_ARMI}
\alias{make_riverfly_ARMI}
\title{make_riverfly_ARMI}
\usage{
make_riverfly_ARMI()
}
\value{
The data frame with riverfly data and an appended ARMI column.
}
\description{
A function to add the ARMI values to the riverfly data
}
