% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_riverfly_ARMI.R
\name{make_riverfly_ARMI}
\alias{make_riverfly_ARMI}
\title{make_riverfly_ARMI}
\usage{
make_riverfly_ARMI(riverfly_data)
}
\arguments{
\item{riverfly_data}{- riverfly data frame from}
}
\value{
The data frame with riverfly data and an appended ARMI column.
}
\description{
A function to add the ARMI values to the riverfly data
}
