% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{plot_breaks}
\alias{plot_breaks}
\title{plot_breaks}
\format{
An object of class \code{data.frame} with 42 rows and 3 columns.
}
\usage{
plot_breaks
}
\description{
plot_breaks
}
\keyword{datasets}
