% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_locations.R
\name{process_locations}
\alias{process_locations}
\title{process_locations}
\usage{
process_locations(
  sampling_locs_url =
    "https://docs.google.com/spreadsheets/d/1ZEkLC3HBkB8SJynA3pHtmntMOiCT8p4e2BFNYsMUR4c/edit?usp=sharing",
  outfall_locs_url =
    "https://docs.google.com/spreadsheets/d/1JJ8bPWppVKbmCfllIevrVmt_dcoswOim7Cos418Ot6w/edit?gid=0#gid=0"
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
