% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_plots.r
\name{species_plots}
\alias{species_plots}
\title{species_plots}
\usage{
species_plots(riverfly_data, sampling_locs)
}
\arguments{
\item{riverfly_data}{The cleaned Riverfly data}

\item{sampling_locs}{The sampling locations for Riverfly data}
}
\value{
a list with four elements: Riverfly_Species_Plot which
has all Riverfly species, Riverfly_Species_Plot_Recent which has
all observations of Riverfly species for the past year,
Other_Species_Plot which has all observations of other species, and
Other_Species_Plot_Recent
}
\description{
Function to create the data for the Riverfly plots
}
